+++ CHANGELOG.md
@@ -8,6 +8,8 @@ CHANGELOG
 - Support publishing and consuming Policy Packs using any runtime
   [#5102](https://github.com/pulumi/pulumi/pull/5102)
 
+- SDK changes to support Python input/output classes
+  [#5033](https://github.com/pulumi/pulumi/pull/5033)
 
 ## 2.8.1 (2020-08-05)